requirements.txt
discord-py-interactions==4.3.4
langchain==0.0.188
openai==0.27.7
tiktoken==0.4.0
aiohttp==3.8.4
pinecone-client==2.2.1
bs4==0.0.1
python-dotenv==1.0.0
pymongo==4.3.3